Move collections extensions to test projects where used
