+++ Tests/Integration/CloudToDeviceMessageSizeLimitShouldAbandonTests.cs
@@ -9,7 +9,6 @@ namespace LoRaWan.Tests.Integration
     using LoRaTools;
     using LoRaTools.LoRaMessage;
     using LoRaTools.Regions;
-    using LoRaTools.Utils;
     using LoRaWan.NetworkServer;
     using LoRaWan.Tests.Common;
     using Microsoft.Azure.Devices.Client;
